requirements.txt
Django==4.2.7.
gunicorn==21.2.0
google-cloud-bigquery==3.27.0